Splits Arguable into two traits, one for providing DSL, one for providing basid support for arguments
Also allows us to decorate a class as supporting arguments without also exposing a DSL
